+++ src/Backends/Amp.php
@@ -4,10 +4,12 @@ namespace Rubix\ML\Backends;
 
 use Amp\Loop;
 use Rubix\ML\Deferred;
+use Amp\Parallel\Worker\Task;
 use Rubix\ML\Other\Helpers\CPU;
 use Amp\Parallel\Worker\DefaultPool;
 use Amp\Parallel\Worker\CallableTask;
 use InvalidArgumentException;
+use Generator;
 use Closure;
 
 use function Amp\call;
@@ -16,9 +18,9 @@ use function Amp\Promise\all;
 /**
  * Amp
  *
- * Amp Parallel is a multiprocessing subsystem that requires no extensions.
- * It uses a non-blocking concurrency framework that implements coroutines
- * using PHP generator functions under the hood.
+ * Amp Parallel is a multiprocessing subsystem that requires no extensions. It uses a
+ * non-blocking concurrency framework that implements coroutines using PHP generator
+ * functions under the hood.
  *
  * @category    Machine Learning
  * @package     Rubix/ML
@@ -34,13 +36,20 @@ class Amp implements Backend
     protected $pool;
 
     /**
-     * The queue of coroutines.
+     * The queue of coroutines to be processed in parallel.
      *
      * @var \Amp\Promise[]
      */
     protected $queue = [
         //
     ];
+
+    /**
+     * The memoized results of the last parallel computation.
+     *
+     * @var mixed[]
+     */
+    protected $results;
 
     /**
      * @param int|null $workers
@@ -77,19 +86,29 @@ class Amp implements Backend
      */
     public function enqueue(Deferred $deferred, ?Closure $after = null) : void
     {
-        $task = new CallableTask([$deferred, 'compute'], []);
+        $task = new CallableTask($deferred, []);
 
-        $coroutine = call(function () use ($task, $after) {
-            $result = yield $this->pool->enqueue($task);
-
-            if ($after) {
-                $after($result);
-            }
-
-            return $result;
-        });
+        $coroutine = call([$this, 'coroutine'], $task, $after);
 
         $this->queue[] = $coroutine;
+    }
+
+    /**
+     * Create a coroutine for a particular task.
+     *
+     * @param \Amp\Parallel\Worker\Task $task
+     * @param \Closure|null $after
+     * @return \Generator<\Amp\Promise>
+     */
+    public function coroutine(Task $task, ?Closure $after = null) : Generator
+    {
+        $result = yield $this->pool->enqueue($task);
+
+        if ($after) {
+            $after($result);
+        }
+
+        return $result;
     }
 
     /**
@@ -99,22 +118,30 @@ class Amp implements Backend
      */
     public function process() : array
     {
-        $results = [];
+        Loop::run([$this, 'gather']);
 
-        Loop::run(function () use (&$results) {
-            $results = yield all($this->queue);
-        });
+        $this->queue = [];
 
-        $this->flush();
-
-        return $results;
+        return $this->results;
     }
 
     /**
-     * Flush the queue.
+     * Gather and memoize the results from the worker pool.
+     *
+     * @return \Generator<\Amp\Promise>
+     */
+    public function gather() : Generator
+    {
+        $this->results = yield all($this->queue);
+    }
+
+    /**
+     * Flush the queue and clear the memoized results.
      */
     public function flush() : void
     {
         $this->queue = [];
+        
+        unset($this->results);
     }
 }