+++ src/NeuralNet/Layers/BatchNorm.php
@@ -259,7 +259,7 @@ class BatchNorm implements Hidden, Parametric
                 . ' backpropagating.');
         }
 
-        $dOut = $prevGradient->compute();
+        $dOut = $prevGradient();
 
         $dBeta = $dOut->sum();
         $dGamma = $dOut->multiply($this->xHat)->sum();