+++ CHANGELOG.md
@@ -8,10 +8,11 @@
     - Missing Data Imputer now passes through other data types
     - Changed order of Missing Data Imputer params
     - Renamed high-level resource type to image
-    - Added comb (n choose k) as a global function
+    - Added comb (n choose k) to global functions
     - Image Vectorizer now has grayscale option
     - Clusterers and Anomaly Detectors return integer predictions
     - Ball Tree now compatible with categorical distance kernels
+    - Parallel Learners using Amp Backend are now persistable
 
 - 0.0.18-beta
     - Now requires PHP 7.2 and above